Extract common functionality into base class
